+++ test/curry.test.js
@@ -7,12 +7,6 @@ describe('curry', () => {
     expect(typeof curriedFn).toBe('function');
     expect(typeof curriedFn(8)).toBe('function');
     expect(curriedFn(8)(8)).toBe(16);
-  });
-
-  test('curry function curry function with two arguments with default argument into single function', () => {
-    const curriedFn = curry((n, a) => a + n, [8]);
-
-    expect(typeof curriedFn).toBe('function');
-    expect(curriedFn(8)).toBe(16);
+    expect(curriedFn(8, 8)).toBe(16);
   });
 });